tests/wpt/tests/css/css-overflow/scroll-markers/scroll-marker-group-size-container-query-root.html

<!DOCTYPE html>
<title>CSS Overflow Test: Size container for :root::scroll-marker-group</title>
<link rel="help" href="https://drafts.csswg.org/css-overflow-5/#scroll-marker-group-property">
<link rel="help" href="https://drafts.csswg.org/css-conditional-5/#container-queries">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>
  :root {
    container-type: inline-size;
    width: 400px;
    scroll-marker-group: before;
    @container (width = 400px) {
      &::scroll-marker-group { --test: pass; }
    }
  }
</style>
<script>
  test(() => {
    assert_equals(getComputedStyle(document.documentElement, "::scroll-marker-group").getPropertyValue("--test"), "pass");
  }, "::scroll-marker-group generates a box inside the root element box and can query it as a size container");
</script>
