LayoutTests/imported/w3c/web-platform-tests/css/css-overflow/scroll-markers/scroll-marker-group-size-container-query-root-expected.txt


FAIL ::scroll-marker-group generates a box inside the root element box and can query it as a size container assert_equals: expected "pass" but got ""
